+++ pkg/auth/oauth2/oauth2.go
@@ -1,0 +1,3 @@
+// Package oauth2 provides the implementation of the OAuth 2.0 specification
+// as defined in RFC6749.
+package oauth2