+++ pkg/auth/oauth2/token.go
@@ -1,0 +1,10 @@
+package oauth2
+
+// TokenResponse represents the response type when querying
+// the token endpoint.
+type TokenResponse struct {
+	AccessToken  string `json:"access_token,omitempty"`
+	TokenType    string `json:"token_type,omitempty"`
+	ExpiresIn    int    `json:"expires_in,omitempty"`
+	RefreshToken string `json:"refresh_token,omitempty"`
+}